requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dateutil==2.8.2
Jinja2==3.1.2 